requirements.txt
aiogram==3.13.0
openai==1.61.0
python-dotenv==1.0.1
asyncpg==0.29.0
httpx==0.27.2